dev > main: refactored models - also added poly conditions as internal settings
