+++ documentation/models/app/Models/PolynomialSetting.md
@@ -1,10 +1,10 @@
-# AppConfigSettings
+# PolynomialSetting
 ## Properties
 
 Name | Type | Description | Notes
 ------------ | ------------- | ------------- | -------------
-**units** | [**Map**](string.md) | Standard units for physical conversion. | [default to null]
-**polynomial** | [**Map**](PolynomialSetting.md) | Conditions for fitting (polynomial) curves to raw. | [default to null]
+**degree** | [**Integer**](integer.md) | Degree of polynomial to be fitted. | [default to null]
+**conditions** | [**List**](PolynomialCondition.md) | Conditions to force on fitted polynomial | [default to null]
 
 [[Back to Model list]](../README.md#documentation-for-models) [[Back to API list]](../README.md#documentation-for-api-endpoints) [[Back to README]](../README.md)
